backend/requirements.txt
# requirements.txt

fastapi
sqlalchemy
faker>=18.0
"uvicorn[standard]"
openai
python-dotenv